# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: SessionLayer Control Plane API
  version: 0.1.0
  summary: Contract-first REST surface for the SessionLayer Control Plane.
  description: |
    The **SessionLayer Control Plane API** is the contract-first source of truth
    for the platform's REST surface (Design §13, FR-API-1). The UI is a *client* of
    this contract; server interfaces and typed clients are code-generated from this
    document, and CI fails if generated code drifts from the spec. Postgres is the
    single source of truth; the UI and the API share one write path (D11).

    **Complete contract.** This document is the complete resource surface: the
    config resources (`node-policies`, `rules`, `roles` + `role-bindings`, `cas`,
    `capability-defs`, `jit-policies`, `breakglass-policies`,
    `service-accounts`, `session-limit-policies`, `operator-settings`), the
    runtime resources (`nodes`, `gateways`, `sessions`, `session-leases`,
    `locks`, `jit-requests`, `join-tokens`, `otp`, device-flow, `pins`,
    `breakglass`, `recordings`, `audit-events`), and the operational probes.

    **Config vs runtime.** Config resources carry an `origin` provenance label
    (`api|ui|default`) and an optimistic-concurrency `version`; runtime resources
    (locks, sessions, issuance, audit) are never provenance-labelled. `locks` is an
    API-only runtime resource (Design §13).

    **Conventions.** URI versioning (`/v1`); cursor pagination on collections;
...
    compatibility policy.
...
servers:
...
security:
  - oidcBearer: []
  - clientCredentials: []
  - mtls: []
paths:
...
  /v1/capability-defs:
    get:
      operationId: listCapabilityDefs
      summary: List capability definitions (admin).
      description: |
        Lists the requestable-capability catalogue (`config.capability_def`; Design
        §12A / D14). Cursor-paginated. Platform-RBAC gated (`settings:write`).
      tags:
        - capability-defs
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A page of capability definitions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapabilityDefPage'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: createCapabilityDef
      summary: Create a capability definition (admin).
      description: |
        Adds a capability to the catalogue. A `name` outside the closed capability
        set is a `422`; a duplicate is a `409`. Platform-RBAC gated (`settings:write`)
        + audited.
      tags:
        - capability-defs
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCapabilityDefRequest'
      responses:
        '201':
          description: The capability definition was created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapabilityDefResource'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
  /v1/capability-defs/{capabilityDefId}:
    parameters:
      - $ref: '#/components/parameters/CapabilityDefId'
    get:
      operationId: getCapabilityDef
      summary: Get a capability definition (admin).
      description: Returns one capability definition by id. Platform-RBAC gated (`settings:write`).
      tags:
        - capability-defs
      responses:
        '200':
          description: The capability definition.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapabilityDefResource'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
    put:
      operationId: updateCapabilityDef
      summary: Update a capability definition (admin).
      description: |
        Updates a capability definition's description (the `name` is immutable). A
        stale `version` is a `409`. Platform-RBAC gated (`settings:write`) + audited.
      tags:
        - capability-defs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCapabilityDefRequest'
      responses:
        '200':
          description: The updated capability definition.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapabilityDefResource'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: deleteCapabilityDef
      summary: Delete a capability definition (admin).
      description: Deletes a capability definition (idempotent). Platform-RBAC gated (`settings:write`) + audited.
      tags:
        - capability-defs
      responses:
        '204':
          description: The capability definition was deleted (idempotent).
        '403':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
...
  /v1/breakglass-policies:
    get:
      operationId: listBreakglassPolicies
      summary: List break-glass policies (admin).
      description: |
        Lists break-glass policies (`config.breakglass_policy`; FR-ACC-6) —
        recording-strict, alert target, review-required, and the IdP-independent
        auth path. Cursor-paginated. Platform-RBAC gated (`breakglass:manage`).
      tags:
        - breakglass-policies
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A page of break-glass policies.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BreakglassPolicyPage'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: createBreakglassPolicy
      summary: Create a break-glass policy (admin).
      description: |
        Creates a break-glass policy. An empty `alertTarget` or a bad `authPath`
        (not fido2|offline_code) is a `422`. Platform-RBAC gated (`breakglass:manage`)
        + audited.
      tags:
        - breakglass-policies
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBreakglassPolicyRequest'
      responses:
        '201':
          description: The break-glass policy was created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BreakglassPolicyResource'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
  /v1/breakglass-policies/{breakglassPolicyId}:
    parameters:
      - $ref: '#/components/parameters/BreakglassPolicyId'
    get:
      operationId: getBreakglassPolicy
      summary: Get a break-glass policy (admin).
      description: Returns one break-glass policy by id. Platform-RBAC gated (`breakglass:manage`).
      tags:
        - breakglass-policies
      responses:
        '200':
          description: The break-glass policy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BreakglassPolicyResource'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
    put:
      operationId: updateBreakglassPolicy
      summary: Update a break-glass policy (admin).
      description: |
        Replaces a break-glass policy's mutable fields (the `name` is immutable). A
        bad `authPath`/empty `alertTarget` is a `422`; a stale `version` is a `409`.
        Platform-RBAC gated (`breakglass:manage`) + audited.
      tags:
        - breakglass-policies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBreakglassPolicyRequest'
      responses:
        '200':
          description: The updated break-glass policy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BreakglassPolicyResource'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: deleteBreakglassPolicy
      summary: Delete a break-glass policy (admin).
      description: Deletes a break-glass policy (idempotent). Platform-RBAC gated (`breakglass:manage`) + audited.
      tags:
        - breakglass-policies
      responses:
        '204':
          description: The break-glass policy was deleted (idempotent).
        '403':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
  # ===================================================================
  # Session-limit policies (FR-SESS-3 write surface).
  # Per-identity overrides for the three session-limit knobs; every stored
  # value is ENFORCED (no dead config): cap + duration at Authorize, idle
  # via the signed decision context at the Gateway.
  # ===================================================================
...
components:
  securitySchemes:
    oidcBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        OIDC bearer access via an ID/JWT token validated by the Control Plane
        (Design §5.7, FR-AUTH-17). The ID token is the authentication proof.
    clientCredentials:
      type: oauth2
      description: |
        OAuth 2.0 client-credentials for machine consumers (Design §5.6,
        FR-AUTH-12). `private_key_jwt`/mTLS client authentication is preferred
        over static secrets.
      flows:
        clientCredentials:
          tokenUrl: https://controlplane.sessionlayer.example/oauth2/token
          scopes: {}
    mtls:
      type: mutualTLS
      description: |
        Mutual TLS client-certificate authentication (Design §5.7, FR-AUTH-17),
        also used by the internal CP<->Gateway plane.
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      description: |
        Opaque forward-only pagination cursor from a previous page's `nextCursor`.
        Omit for the first page. An unrecognised cursor is a `400`.
      schema:
        type: string
        maxLength: 512
    Limit:
      name: limit
      in: query
      required: false
      description: Maximum items to return in one page (the server clamps to its maximum).
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 200
        default: 50
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Optional client-generated idempotency key scoping a mutating request. A
        retry with the same key, method, and path returns the original response and
        never repeats the side effect; the same key with a different request body is
        a `422`. Keys are retained for a bounded TTL.
      schema:
        type: string
        minLength: 1
        maxLength: 255
...
    CapabilityDefId:
      name: capabilityDefId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    BreakglassPolicyId:
      name: breakglassPolicyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Problem:
      description: An error occurred. The body is an RFC 9457 problem document.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  schemas:
    ProblemDetails:
      type: object
      title: RFC 9457 Problem Details
      description: |
        Problem details for HTTP APIs as defined by RFC 9457 (obsoletes RFC 7807).
        Error responses use media type `application/problem+json`.
      properties:
        type:
          type: string
          format: uri-reference
          default: about:blank
          description: A URI reference identifying the problem type.
        title:
          type: string
          description: A short, human-readable summary of the problem type.
        status:
          type: integer
          format: int32
          minimum: 100
          maximum: 599
          description: The HTTP status code.
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence.
        instance:
          type: string
          format: uri-reference
          description: A URI reference identifying the specific occurrence.
      additionalProperties: true
...
    Origin:
      type: string
      title: Config Provenance
      description: |
        Provenance of a config row (Design §13). `api`/`ui` mark which admin surface
        last wrote it; `default` is a seeded/cold-start default. Read-only.
      enum:
        - api
        - ui
        - default
    Capability:
      type: string
      title: SSH Capability
      description: A policy-gated SSH channel capability (Design D14).
      enum:
        - shell
        - exec
        - sftp
        - scp
        - port_forward_local
        - port_forward_remote
        - agent_forward
        - x11
...
    BreakglassAuthPath:
      type: string
      title: Break-glass Auth Path
      enum:
        - fido2
        - offline_code
...
    CapabilityDefResource:
      type: object
      title: Capability Definition
      required:
        - id
        - name
        - origin
        - version
      properties:
        id:
          type: string
          format: uuid
        name:
          $ref: '#/components/schemas/Capability'
        description:
          type: string
        origin:
          $ref: '#/components/schemas/Origin'
        version:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: false
    CreateCapabilityDefRequest:
      type: object
      title: Create Capability Definition Request
      required:
        - name
      properties:
        name:
          $ref: '#/components/schemas/Capability'
        description:
          type: string
      additionalProperties: false
    UpdateCapabilityDefRequest:
      type: object
      title: Update Capability Definition Request
      required:
        - version
      properties:
        description:
          type: string
        version:
          type: integer
          format: int64
      additionalProperties: false
    CapabilityDefPage:
      type: object
      title: Capability Definition Page
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CapabilityDefResource'
        nextCursor:
          type: string
      additionalProperties: false
    # ============ jit-policies (jit_policy) ============
...
    BreakglassPolicyResource:
      type: object
      title: Break-glass Policy
      required:
        - id
        - name
        - recordingStrict
        - alertTarget
        - reviewRequired
        - authPath
        - origin
        - version
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        recordingStrict:
          type: boolean
        alertTarget:
          type: string
        reviewRequired:
          type: boolean
        authPath:
          $ref: '#/components/schemas/BreakglassAuthPath'
        origin:
          $ref: '#/components/schemas/Origin'
        version:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: false
    CreateBreakglassPolicyRequest:
      type: object
      title: Create Break-glass Policy Request
      required:
        - name
        - alertTarget
      properties:
        name:
          type: string
          minLength: 1
        recordingStrict:
          type: boolean
        alertTarget:
          type: string
          minLength: 1
        reviewRequired:
          type: boolean
        authPath:
          $ref: '#/components/schemas/BreakglassAuthPath'
      additionalProperties: false
    UpdateBreakglassPolicyRequest:
      type: object
      title: Update Break-glass Policy Request
      required:
        - alertTarget
        - version
      properties:
        recordingStrict:
          type: boolean
        alertTarget:
          type: string
          minLength: 1
        reviewRequired:
          type: boolean
        authPath:
          $ref: '#/components/schemas/BreakglassAuthPath'
        version:
          type: integer
          format: int64
      additionalProperties: false
    BreakglassPolicyPage:
      type: object
      title: Break-glass Policy Page
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/BreakglassPolicyResource'
        nextCursor:
          type: string
      additionalProperties: false
    # ============ session-limit-policies (session_limit_policy) ============
...